split the renderAttractions into the details.js file and renderAttractions file
